mlops/requirements.txt
abnf==2.2.0
about-time==4.2.1
absl-py==2.4.0
alive-progress==3.3.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-generator==1.10
async-lru==2.1.0
attrs==25.4.0
babel==2.18.0
backoff==2.2.1
beautifulsoup4==4.14.3
black==26.1.0
bleach==6.3.0
cachetools==7.0.1
category-encoders==2.9.0
certifi==2026.1.4
cffi==2.0.0
cfgv==3.5.0
chardet==6.0.0.post1
charset-normalizer==3.4.4
cint==1.0.0
click==8.3.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.5
curl-cffi==0.13.0
cycler==0.12.1
datetime==6.0
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
deprecated==1.3.1
diskcache==5.6.3
distlib==0.4.0
eli5==0.16.0
et-xmlfile==2.0.0
executing==2.2.1
fastapi==0.128.0
fastjsonschema==2.21.2
fickling==0.1.8
filelock==3.20.3
fire==0.7.1
flatbuffers==25.12.19
fonttools==4.61.1
fqdn==1.5.1
frozendict==2.4.7
gdown==4.7.1
gensim==4.4.0
gitdb==4.0.12
gitpython==3.1.46
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
googleapis-common-protos==1.72.0
gql==4.0.0
graphemeu==0.7.2
graphql-core==3.2.7
graphviz==0.21
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
icecream==2.1.10
identify==2.6.16
idna==3.11
importlib-metadata==8.7.1
intervaltree==3.2.1
ipykernel==7.1.0
ipython==9.10.0
ipython-pygments-lexers==1.1.1
isoduration==20.11.0
isort==8.0.0
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.3
json5==0.13.0
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-client==8.8.0
jupyter-core==5.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter-server==2.17.0
jupyter-server-terminals==0.5.4
jupyterlab==4.5.3
jupyterlab-pygments==0.3.0
jupyterlab-server==2.28.0
kaggle==1.8.4
kagglesdk==0.1.15
kaitaistruct==0.11
kiwisolver==1.4.9
lark==1.3.1
lightgbm==4.6.0
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mediapipe==0.10.32
mistune==3.2.0
mlxtend==0.24.0
mplfinance==0.12.10b0
multidict==6.7.1
multipledispatch==1.0.0
multitasking==0.0.12
mypy-extensions==1.1.0
narwhals==2.15.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
nodeenv==1.10.0
notebook==7.5.3
notebook-shim==0.2.4
numpy==2.4.2
opencv-contrib-python==4.13.0.90
opencv-python==4.13.0.90
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==26.0
pandas==3.0.0
pandas-stubs==3.0.0.260204
pandocfilters==1.5.1
parso==0.8.5
pathspec==1.0.4
patsy==1.0.2
pdfminer-six==20260107
peewee==3.19.0
pexpect==4.9.0
pillow==12.1.0
platformdirs==4.5.1
plotly==6.5.2
polyfile-weave==0.5.9
pre-commit==4.5.1
prometheus-client==0.24.1
prompt-toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.2.2
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1-modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pydotplus==2.0.2
pygame==2.6.1
pygments==2.19.2
pykrx==1.0.51
pyparsing==3.3.2
pysnooper==1.2.3
pysocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-slugify==8.0.4
pytokens==0.4.1
pytz==2025.2
pyyaml==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.3.2
rpds-py==0.30.0
rsa==4.9.1
schedule==1.2.2
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
selenium==4.40.0
send2trash==2.1.0
sentry-sdk==2.53.0
setuptools==80.10.2
six==1.17.0
smart-open==7.5.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.5
soupsieve==2.8.3
sqlalchemy==2.0.47
stack-data==0.6.3
starlette==0.50.0
statsmodels==0.14.6
tabulate==0.9.0
tenacity==9.1.4
termcolor==3.3.0
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.6.0
tinycss2==1.4.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
trio==0.32.0
trio-typing==0.10.0
trio-websocket==0.12.2
types-certifi==2021.10.8.3
types-requests==2.32.4.20260107
types-urllib3==1.26.25.14
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.40.0
virtualenv==20.36.1
wandb==0.25.0
wcwidth==0.5.3
weave==0.52.28
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
wrapt==2.1.0
wsproto==1.3.2
xgboost==3.1.3
xlrd==2.0.2
yarl==1.22.0
yfinance==1.1.0
zipp==3.23.0
zope-interface==8.2
